make __exec_table_align no-op
it turned out that this was not what was making the performance
differences in question.

the real trigger was the "-dirty" suffix added to the version string.
(it was added because i tested the change without committing.)
adding 6 letters to other string literal seems to cause the same
performance "improvement" as well.
although i'm still not sure how cstring changes can cause the performance
differences, it happens regardless of the alignment of this table.