add an example program to demonstrate module_writer.c
the original intention of module_writer.c was to create
a kind of aot compiler based on it. but the plan didn't happen
to get tractions and the writer functionality has always been unused.
now i'm feeling that it's better to introduce a simple way
to test it to avoid regressions.